+++ src/gf16.h
@@ -24,32 +24,32 @@
 
 uint8_t inv16(uint8_t element);
 
-void maddrc16_imul_scalar(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc16_imul_gpr32(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc16_imul_gpr64(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc16_flat_table(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc16_log_table(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
+void maddrc16_imul_scalar(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc16_imul_gpr32(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc16_imul_gpr64(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc16_flat_table(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc16_log_table(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
 
-void mulrc16_imul_scalar(uint8_t *region, uint8_t constant, int length);
-void mulrc16_imul_gpr32(uint8_t *region, uint8_t constant, int length);
-void mulrc16_imul_gpr64(uint8_t *region, uint8_t constant, int length);
+void mulrc16_imul_scalar(uint8_t *region, uint8_t constant, size_t length);
+void mulrc16_imul_gpr32(uint8_t *region, uint8_t constant, size_t length);
+void mulrc16_imul_gpr64(uint8_t *region, uint8_t constant, size_t length);
 
 #ifdef __x86_64__
-void maddrc16_imul_sse2(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc16_imul_avx2(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc16_shuffle_ssse3(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc16_shuffle_avx2(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
+void maddrc16_imul_sse2(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc16_imul_avx2(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc16_shuffle_ssse3(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc16_shuffle_avx2(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
 
-void mulrc16_imul_sse2(uint8_t *region, uint8_t constant, int length);
-void mulrc16_shuffle_ssse3(uint8_t *region, uint8_t constant, int length);
-void mulrc16_shuffle_avx2(uint8_t *region, uint8_t constant, int length);
+void mulrc16_imul_sse2(uint8_t *region, uint8_t constant, size_t length);
+void mulrc16_shuffle_ssse3(uint8_t *region, uint8_t constant, size_t length);
+void mulrc16_shuffle_avx2(uint8_t *region, uint8_t constant, size_t length);
 #endif
 
 #ifdef __arm__
-void maddrc16_shuffle_neon(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc16_imul_neon_64(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc16_imul_neon_128(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void mulrc16_shuffle_neon(uint8_t *region, uint8_t constant, int length);
+void maddrc16_shuffle_neon(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc16_imul_neon_64(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc16_imul_neon_128(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void mulrc16_shuffle_neon(uint8_t *region, uint8_t constant, size_t length);
 #endif
 
 #endif
