+++ src/gf256.h
@@ -24,33 +24,33 @@
 
 uint8_t inv256(uint8_t element);
 
-void maddrc256_pdiv(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc256_log_table(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc256_flat_table(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc256_imul_gpr32(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc256_imul_gpr64(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
+void maddrc256_pdiv(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc256_log_table(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc256_flat_table(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc256_imul_gpr32(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc256_imul_gpr64(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
 
-void mulrc256_pdiv(uint8_t *region, uint8_t constant, int length);
-void mulrc256_imul_gpr32(uint8_t *region, uint8_t constant, int length);
-void mulrc256_imul_gpr64(uint8_t *region, uint8_t constant, int length);
+void mulrc256_pdiv(uint8_t *region, uint8_t constant, size_t length);
+void mulrc256_imul_gpr32(uint8_t *region, uint8_t constant, size_t length);
+void mulrc256_imul_gpr64(uint8_t *region, uint8_t constant, size_t length);
 
 #ifdef __x86_64__
-void maddrc256_imul_sse2(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc256_imul_avx2(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc256_shuffle_ssse3(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc256_shuffle_avx2(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
+void maddrc256_imul_sse2(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc256_imul_avx2(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc256_shuffle_ssse3(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc256_shuffle_avx2(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
 
-void mulrc256_imul_sse2(uint8_t *region, uint8_t constant, int length);
-void mulrc256_shuffle_ssse3(uint8_t *region, uint8_t constant, int length);
-void mulrc256_shuffle_avx2(uint8_t *region, uint8_t constant, int length);
+void mulrc256_imul_sse2(uint8_t *region, uint8_t constant, size_t length);
+void mulrc256_shuffle_ssse3(uint8_t *region, uint8_t constant, size_t length);
+void mulrc256_shuffle_avx2(uint8_t *region, uint8_t constant, size_t length);
 #endif
 
 #ifdef __arm__
-void maddrc256_shuffle_neon(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc256_imul_neon_64(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
-void maddrc256_imul_neon_128(uint8_t *region1, const uint8_t *region2, uint8_t constant, int length);
+void maddrc256_shuffle_neon(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc256_imul_neon_64(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
+void maddrc256_imul_neon_128(uint8_t *region1, const uint8_t *region2, uint8_t constant, size_t length);
 
-void mulrc256_shuffle_neon(uint8_t *region, uint8_t constant, int length);
+void mulrc256_shuffle_neon(uint8_t *region, uint8_t constant, size_t length);
 #endif
 
 #endif
